+++ tests/unit/apps/grievance/test_model_validation.py
@@ -5,13 +5,13 @@ from hct_mis_api.apps.account.fixtures import UserFactory
 from hct_mis_api.apps.core.fixtures import create_afghanistan
 from hct_mis_api.apps.core.models import BusinessArea
 from hct_mis_api.apps.grievance.models import GrievanceTicket
+from hct_mis_api.apps.payment.admin import FspXlsxTemplatePerDeliveryMechanismForm
 from hct_mis_api.apps.payment.fixtures import (
     FinancialServiceProviderFactory,
     FinancialServiceProviderXlsxTemplateFactory,
-    FspXlsxTemplatePerDeliveryMechanismFactory,
     generate_delivery_mechanisms,
 )
-from hct_mis_api.apps.payment.models import DeliveryMechanism
+from hct_mis_api.apps.payment.models import DeliveryMechanism, FinancialServiceProvider
 
 
 class TestGrievanceModelValidation(TestCase):
@@ -84,38 +84,77 @@ class TestFspXlsxTemplatePerDeliveryMechanismValidation(TestCase):
         create_afghanistan()
         cls.user = UserFactory.create()
         generate_delivery_mechanisms()
+        cls.dm_transfer_to_account = DeliveryMechanism.objects.get(code="transfer_to_account")
 
-    def test_clean(self) -> None:
-        dm_cash = DeliveryMechanism.objects.get(code="cash")
-        dm_atm_card = DeliveryMechanism.objects.get(code="atm_card")
-        fsp = FinancialServiceProviderFactory()
-        fsp.delivery_mechanisms.set([dm_atm_card])
-        template = FinancialServiceProviderXlsxTemplateFactory()
-        template_per_dm_cash = FspXlsxTemplatePerDeliveryMechanismFactory(
-            financial_service_provider=fsp,
-            delivery_mechanism=dm_cash,
-            xlsx_template=template,
+    def test_admin_form_clean(self) -> None:
+        fsp_xls_template = FinancialServiceProviderXlsxTemplateFactory(
+            core_fields=["bank_name__transfer_to_account", "bank_account_number__transfer_to_account"]
         )
 
-        with self.assertRaisesMessage(
-            ValidationError,
-            f"Delivery Mechanism {template_per_dm_cash.delivery_mechanism} is not supported by Financial Service Provider {fsp}",
-        ):
-            template_per_dm_cash.clean()
-
-        template_per_dm_atm_card = FspXlsxTemplatePerDeliveryMechanismFactory(
-            financial_service_provider=fsp,
-            delivery_mechanism=dm_atm_card,
-            xlsx_template=template,
+        fsp = FinancialServiceProviderFactory(
+            name="Test FSP",
+            vision_vendor_number="123",
+            communication_channel=FinancialServiceProvider.COMMUNICATION_CHANNEL_API,
         )
+        fsp.delivery_mechanisms.add(self.dm_transfer_to_account)
 
+        # test valid form
+        form_data_standalone = {
+            "financial_service_provider": fsp.id,
+            "delivery_mechanism": self.dm_transfer_to_account.id,
+            "xlsx_template": fsp_xls_template.id,
+        }
+        form = FspXlsxTemplatePerDeliveryMechanismForm(data=form_data_standalone)
+        self.assertTrue(form.is_valid())
+        form.clean()
+
+        # test inline form data valid
+        form_data_inline = {
+            "financial_service_provider": fsp.id,
+            "delivery_mechanism": self.dm_transfer_to_account.id,
+            "xlsx_template": fsp_xls_template.id,
+            "delivery_mechanisms": [str(self.dm_transfer_to_account.id)],
+        }
+        form = FspXlsxTemplatePerDeliveryMechanismForm(data=form_data_inline)
+        self.assertTrue(form.is_valid())
+        form.clean()
+
+        # test missing required core fields
+        fsp_xls_template.core_fields = []
+        fsp_xls_template.save()
+
+        form = FspXlsxTemplatePerDeliveryMechanismForm(data=form_data_standalone)
+        self.assertFalse(form.is_valid())
         with self.assertRaisesMessage(
             ValidationError,
-            f"['card_number__atm_card', 'card_expiry_date__atm_card', 'name_of_cardholder__atm_card'] fields are required by delivery mechanism "
-            f"{template_per_dm_atm_card.delivery_mechanism} and must be present in the template core fields",
+            "[\"['bank_name__transfer_to_account', 'bank_account_number__transfer_to_account'] fields are required by delivery mechanism Transfer to Account and must be present in the template core fields\"]",
         ):
-            template_per_dm_atm_card.clean()
+            form.clean()
 
-        template.core_fields = ["card_number__atm_card", "card_expiry_date__atm_card", "name_of_cardholder__atm_card"]
-        template.save()
-        template_per_dm_atm_card.clean()
+        fsp_xls_template.core_fields = ["bank_name__transfer_to_account", "bank_account_number__transfer_to_account"]
+        fsp_xls_template.save()
+
+        # test delivery mechanism not supported
+        fsp.delivery_mechanisms.remove(self.dm_transfer_to_account)
+        form = FspXlsxTemplatePerDeliveryMechanismForm(data=form_data_standalone)
+        self.assertFalse(form.is_valid())
+        with self.assertRaisesMessage(
+            ValidationError,
+            "['Delivery Mechanism Transfer to Account is not supported by Financial Service Provider Test FSP (123): API']",
+        ):
+            form.clean()
+
+        # test inline form data invalid
+        form_data_inline = {
+            "financial_service_provider": fsp.id,
+            "delivery_mechanism": self.dm_transfer_to_account.id,
+            "xlsx_template": fsp_xls_template.id,
+            "delivery_mechanisms": ["12313213123"],
+        }
+        form = FspXlsxTemplatePerDeliveryMechanismForm(data=form_data_inline)
+        self.assertFalse(form.is_valid())
+        with self.assertRaisesMessage(
+            ValidationError,
+            "['Delivery Mechanism Transfer to Account is not supported by Financial Service Provider Test FSP (123): API']",
+        ):
+            form.clean()
