+++ tests/unit/apps/payment/test_models1.py
@@ -4,7 +4,6 @@ from typing import Any
 from unittest.mock import MagicMock, patch
 
 from django import forms
-from django.core.exceptions import ValidationError
 from django.db import models
 from django.db.utils import IntegrityError
 from django.test import TestCase
@@ -22,20 +21,16 @@ from hct_mis_api.apps.household.fixtures import (
     IndividualFactory,
     create_household,
 )
-from hct_mis_api.apps.payment.admin import FspXlsxTemplatePerDeliveryMechanismForm
 from hct_mis_api.apps.payment.fields import DynamicChoiceArrayField, DynamicChoiceField
 from hct_mis_api.apps.payment.fixtures import (
     DeliveryMechanismPerPaymentPlanFactory,
     FinancialServiceProviderFactory,
-    FinancialServiceProviderXlsxTemplateFactory,
     PaymentFactory,
     PaymentPlanFactory,
     PaymentPlanSplitFactory,
     RealProgramFactory,
-    generate_delivery_mechanisms,
 )
 from hct_mis_api.apps.payment.models import (
-    DeliveryMechanism,
     FinancialServiceProvider,
     FinancialServiceProviderXlsxTemplate,
     Payment,
@@ -398,8 +393,6 @@ class TestFinancialServiceProviderModel(TestCase):
     def setUpTestData(cls) -> None:
         super().setUpTestData()
         create_afghanistan()
-        generate_delivery_mechanisms()
-        cls.dm_transfer_to_account = DeliveryMechanism.objects.get(code="transfer_to_account")
         cls.business_area = BusinessArea.objects.get(slug="afghanistan")
 
     def test_properties(self) -> None:
@@ -498,79 +491,6 @@ class TestFinancialServiceProviderModel(TestCase):
         self.assertEqual(blockchain_name, individuals[0].blockchain_name)
         wallet_address = fsp_xlsx_template.get_column_from_core_field(payment, "wallet_address")
         self.assertEqual(wallet_address, individuals[0].wallet_address)
-
-    def test_admin_form_clean(self) -> None:
-        fsp_xls_template = FinancialServiceProviderXlsxTemplateFactory(
-            core_fields=["bank_name__transfer_to_account", "bank_account_number__transfer_to_account"]
-        )
-
-        fsp = FinancialServiceProviderFactory(
-            name="Test FSP",
-            vision_vendor_number="123",
-            communication_channel=FinancialServiceProvider.COMMUNICATION_CHANNEL_API,
-        )
-        fsp.delivery_mechanisms.add(self.dm_transfer_to_account)
-
-        # test valid form
-        form_data_standalone = {
-            "financial_service_provider": fsp.id,
-            "delivery_mechanism": self.dm_transfer_to_account.id,
-            "xlsx_template": fsp_xls_template.id,
-        }
-        form = FspXlsxTemplatePerDeliveryMechanismForm(data=form_data_standalone)
-        self.assertTrue(form.is_valid())
-        form.clean()
-
-        # test inline form data valid
-        form_data_inline = {
-            "financial_service_provider": fsp.id,
-            "delivery_mechanism": self.dm_transfer_to_account.id,
-            "xlsx_template": fsp_xls_template.id,
-            "delivery_mechanisms": [str(self.dm_transfer_to_account.id)],
-        }
-        form = FspXlsxTemplatePerDeliveryMechanismForm(data=form_data_inline)
-        self.assertTrue(form.is_valid())
-        form.clean()
-
-        # test missing required core fields
-        fsp_xls_template.core_fields = []
-        fsp_xls_template.save()
-
-        form = FspXlsxTemplatePerDeliveryMechanismForm(data=form_data_standalone)
-        self.assertFalse(form.is_valid())
-        with self.assertRaisesMessage(
-            ValidationError,
-            "[\"['bank_name__transfer_to_account', 'bank_account_number__transfer_to_account'] fields are required by delivery mechanism Transfer to Account and must be present in the template core fields\"]",
-        ):
-            form.clean()
-
-        fsp_xls_template.core_fields = ["bank_name__transfer_to_account", "bank_account_number__transfer_to_account"]
-        fsp_xls_template.save()
-
-        # test delivery mechanism not supported
-        fsp.delivery_mechanisms.remove(self.dm_transfer_to_account)
-        form = FspXlsxTemplatePerDeliveryMechanismForm(data=form_data_standalone)
-        self.assertFalse(form.is_valid())
-        with self.assertRaisesMessage(
-            ValidationError,
-            "['Delivery Mechanism Transfer to Account is not supported by Financial Service Provider Test FSP (123): API']",
-        ):
-            form.clean()
-
-        # test inline form data invalid
-        form_data_inline = {
-            "financial_service_provider": fsp.id,
-            "delivery_mechanism": self.dm_transfer_to_account.id,
-            "xlsx_template": fsp_xls_template.id,
-            "delivery_mechanisms": ["12313213123"],
-        }
-        form = FspXlsxTemplatePerDeliveryMechanismForm(data=form_data_inline)
-        self.assertFalse(form.is_valid())
-        with self.assertRaisesMessage(
-            ValidationError,
-            "['Delivery Mechanism Transfer to Account is not supported by Financial Service Provider Test FSP (123): API']",
-        ):
-            form.clean()
 
 
 class TestDynamicChoiceArrayField(TestCase):
